requirements.txt
# Core Dependencies
streamlit>=1.33.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-google-genai>=0.0.5
langchain-ollama>=0.1.0
langchain-community>=0.0.20

# Environment & Configuration
python-dotenv>=1.0.0

# Enhanced UI Components
streamlit-chat>=0.1.1

# Utilities
pydantic>=2.0.0

# File Processing (Optional - for file attachments)
PyPDF2>=3.0.0
python-docx>=1.0.0

# Speech Recognition (Optional - for voice input)
SpeechRecognition>=3.10.0
